requirements.txt
flask
flask-cors
requests
python-dotenv
loguru
pydantic>=2.0
flask
flask-cors
loguru
pydantic
python-dotenv
requests
gunicorn
